Streaming/requirements.txt
alabaster==0.7.8
anaconda-client==1.4.0
anaconda-navigator==1.2.1
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.2.1
awscli==1.11.47
awsebcli==3.10.0
Babel==2.3.3
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
bitarray==0.8.1
blessed==1.14.1
bokeh==0.12.0
boto==2.40.0
boto3==1.4.4
botocore==1.5.24
Bottleneck==1.1.0
cdecimal==2.3
cement==2.8.2
certifi==2017.1.23
cffi==1.6.0
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.2.13
conda-build==1.21.3
configobj==5.0.6
configparser==3.5.0b2
contextlib2==0.5.3
cryptography==1.3.1
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
decorator==4.0.10
dill==0.2.5
Django==1.8.13
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
dynd==0.7.3.dev1
elasticsearch==1.9.0
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.11.1
Flask-Cors==2.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gevent==1.1.1
greenlet==0.4.10
grin==1.2.1
h5py==2.6.0
HeapDict==1.0.0
idna==2.1
imagesize==0.7.1
ipaddress==1.0.14
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jmespath==0.9.2
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
kazoo==2.2.1
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.9.1
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.11.1
oauthlib==2.0.1
openpyxl==2.3.2
packaging==16.8
pandas==0.18.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
pathspec==0.5.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ply==3.8
psutil==4.3.0
ptyprocess==0.5.1
PuLP==1.6.1
py==1.4.31
pyasn1==0.1.9
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.2.3
Pygments==2.1.3
pykafka==2.5.0
pymongo==3.3.1
pyOpenSSL==16.0.0
pyparsing==2.2.0
pysolr==3.6.0
pytest==2.9.2
python-dateutil==2.6.0
pytz==2016.4
PyYAML==3.12
pyzmq==15.2.0
qtconsole==4.2.1
QtPy==1.0.2
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rope==0.9.4
rsa==3.4.2
ruamel-yaml===-VERSION
s3transfer==0.1.10
scikit-learn==0.17.1
scipy==0.17.1
semantic-version==2.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
spyder==2.3.9
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
tabulate==0.7.5
termcolor==1.1.0
terminado==0.6
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
tweepy==3.5.0
unicodecsv==0.14.1
urllib3==1.20
virtualenv==15.1.0
watson-developer-cloud==0.25.2
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.11.10
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
